Split scoring/weights.rs into focused sub-modules (#609)
Split weights.rs (~37KB) into weights/ directory with:
- mod.rs: public API, re-exports, constants
- calculation.rs: core weight calculation logic
- normalisation.rs: range-aware weight computation
- adjustment.rs: dynamic weight adjustments

Public API unchanged. All tests pass.
